Feature/promo code

+++ WebProjekat2022/webProjekat2022/src/beans/Customer.java
@@ -3,25 +3,25 @@ package beans;
 import java.io.Serializable;
 
 public class Customer extends User implements Serializable {
-    private long points;
+    private double points;
     private String customerType;
     private String membership;
 
     public Customer() {
     }
 
-    public Customer(String username, String password, String name, String lastName, Gender gender, String dateOfBirth, Role role, long points, String customerType, String membership) {
+    public Customer(String username, String password, String name, String lastName, Gender gender, String dateOfBirth, Role role, double points, String customerType, String membership) {
         super(username, password, name, lastName, gender, dateOfBirth, role);
         this.points = points;
         this.customerType = customerType;
         this.membership = membership;
     }
 
-    public long getPoints() {
+    public double getPoints() {
         return points;
     }
 
-    public void setPoints(long points) {
+    public void setPoints(double points) {
         this.points = points;
     }
